Tues, 11.14 회원 탈퇴시 권한 검증

+++ src/main/java/com/example/prj1be20231109/controller/MemberController.java
@@ -73,9 +73,15 @@ public class MemberController {
     }
 
     @DeleteMapping
-    public ResponseEntity delete(String id) {
-        // TODO : 로그인 했는 지? -> 안했으면 401
-        // TODO : 자기 정보인지? -> 아니면 403
+    public ResponseEntity delete(String id,
+                                 @SessionAttribute(value = "login", required = false) Member login) {
+        if (login == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build(); // 401
+        }
+
+        if (!service.hasAccess(id, login)) {
+            return ResponseEntity.status(HttpStatus.FORBIDDEN).build(); // 403
+        }
 
         if (service.deleteMember(id)) {
             return ResponseEntity.ok().build();
